Update Xtestmerger.R
The check that both test and X will have matching levels for factor columns breaks if the levels are the same but in the test some level is not present in the data (which I can imagine happens a lot since the test set is always much smaller). By removing the droplevels function now we really compare the levels on the factors and not the appearance of levels in the data.