Update version
Update version for last pull request. This is relevant for people installing the package from the github repo (since the latest update is not in a CRAN package). If install from github then the version will be higher than the one in the CRAN repository hence it will keep the latest version and not override the old version from the CRAN. Once CRAN is up to date it will just work installing the newest version.